Split fields file into multiple files
