+++ src/datumaro/v2/__init__.py
@@ -3,33 +3,15 @@
 # SPDX-License-Identifier: MIT
 
 from . import converters  # Import converters to register them
-from src.datumaro.v2.converters.converter_registry import ConverterRegistry, converter, find_conversion_path
+from datumaro.v2.converters.converter_registry import ConverterRegistry, converter, find_conversion_path
 from .dataset import Dataset, Sample
 from .export_import import export_dataset, import_dataset
-from .fields import (
-    BBoxField,
-    ImageCallableField,
-    ImageField,
-    ImageInfo,
-    ImageInfoField,
-    ImagePathField,
-    LabelField,
-    PolygonField,
-    RotatedBBoxField,
-    ScoreField,
-    TensorField,
-    bbox_field,
-    image_callable_field,
-    image_field,
-    image_info_field,
-    image_path_field,
-    label_field,
-    polygon_field,
-    rotated_bbox_field,
-    score_field,
-    tensor_field,
-)
-from .schema import AttributeInfo, Field, Schema, Semantic
+from .fields.images import TensorField, tensor_field, ImageField, image_field, ImageInfo, ImageInfoField, \
+    image_info_field, ImagePathField, image_path_field, ImageCallableField, image_callable_field
+from .fields.annotations import BBoxField, bbox_field, RotatedBBoxField, rotated_bbox_field, LabelField, label_field, \
+    ScoreField, score_field, PolygonField, polygon_field
+from .schema import AttributeInfo, Schema
+from .fields import Semantic, Field
 
 # Import tilers and converters implementations to register them
 from .tiling import tilers
